Refactor patreon to Features.
