+++ backend/Veloci.Logic/Bot/Discord/DiscordMessageEventHandler.cs
@@ -3,6 +3,7 @@ using MediatR;
 using Serilog;
 using Veloci.Data.Domain;
 using Veloci.Data.Repositories;
+using Veloci.Logic.Features.Patreon.Notifications;
 using Veloci.Logic.Helpers;
 using Veloci.Logic.Notifications;
 using Veloci.Logic.Services;
