+++ backend/Veloci.Logic/Bot/Telegram/TelegramMessageEventHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Hangfire;
 using MediatR;
+using Veloci.Logic.Features.Patreon.Notifications;
 using Veloci.Logic.Helpers;
 using Veloci.Logic.Notifications;
 using Veloci.Logic.Services;
