+++ backend/Veloci.Logic/Features/Patreon/Models/PatreonOptions.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Veloci.Logic.API.Options;
+namespace Veloci.Logic.Features.Patreon.Models;
 
 public class PatreonOptions
 {
