+++ backend/Veloci.Logic/Features/Patreon/Notifications/MonthlyPatreonSupportersNotification.cs
@@ -1,6 +1,6 @@
 using MediatR;
 using Veloci.Data.Domain;
 
-namespace Veloci.Logic.Notifications;
+namespace Veloci.Logic.Features.Patreon.Notifications;
 
 public record MonthlyPatreonSupportersNotification(List<PatreonSupporter> Supporters) : INotification;
