+++ backend/Veloci.Logic/Features/Patreon/Services/PatreonApiClient.cs
@@ -2,9 +2,9 @@ using System.Net;
 using System.Net.Http.Headers;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
-using Veloci.Logic.Services.Patreon.Models;
+using Veloci.Logic.Features.Patreon.Models;
 
-namespace Veloci.Logic.Services;
+namespace Veloci.Logic.Features.Patreon.Services;
 
 public class PatreonApiClient : IPatreonApiClient
 {
