+++ backend/Veloci.Logic/Features/Patreon/Services/PatreonService.cs
@@ -1,8 +1,8 @@
 using Microsoft.Extensions.Logging;
 using Veloci.Data.Domain;
-using Veloci.Logic.Services.Patreon.Models;
+using Veloci.Logic.Features.Patreon.Models;
 
-namespace Veloci.Logic.Services;
+namespace Veloci.Logic.Features.Patreon.Services;
 
 public class PatreonService : IPatreonService
 {
