+++ backend/Veloci.Logic/Features/Patreon/Views/Patreon/Tokens.cshtml
@@ -1,4 +1,4 @@
-@model Veloci.Web.Controllers.PatreonTokenResponse
+@model Veloci.Logic.Features.Patreon.Models.PatreonTokenResponse
 @{
     ViewData["Title"] = "Patreon OAuth Tokens";
 }
